requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
pymupdf
ollama
tqdm
langchain-text-splitters